Create AllMods
this one is 9 teleport mods, some are advanced some arnt. theese will not work if you just copy the code and put it in a server. and no there arnt deviders so to find the start and finish to a mod use "CTRL F" and search for "    [Info(  " you will find the start of a mod threre.